Remove tons of unneeded address space start subtracts
The chip-select base address must be set according to the size of the corresponding chip-select signals of the group. For example, if CSA1 and CSA0 are each assigned a 2 Mbyte memory space, the CSGBA register must be set in a 4 Mbyte space boundary, such as system address 0 × 0, 0 × 4 Mbyte, 0 × 8 Mbyte, and so on. It cannot be set at 0 × 1 Mbyte, 0 × 2 Mbyte, 0 × 3 Mbyte, 0 × 5 Mbyte, and so on.

M68VZ328 Datasheet